requirements.txt
# FaceMatch Pro - Dependencies

# Core
streamlit>=1.30.0
python>=3.8

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
mtcnn>=0.1.1
Pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Telegram Bot (for feedback)
python-telegram-bot>=20.0
requests>=2.31.0

# Streamlit Extensions
streamlit-option-menu>=0.3.6
streamlit-lottie>=0.0.5
streamlit-camera-input-live>=0.2.0

# Heatmap Visualization
torch-cam>=0.3.2
scikit-learn>=1.3.0

# PDF Generation
reportlab>=4.0.0
fpdf>=1.7.2

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.2

huggingface_hub >= 0.32.4

# Optional: For better performance
# accelerate>=0.24.0
# bitsandbytes>=0.41.0

